backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# IMPORTANT: Core dependencies that uvicorn needs
click==8.1.7
h11==0.14.0
httptools==0.6.1
python-dotenv==1.0.0
pyyaml==6.0.1
watchfiles==0.21.0
websockets==12.0
uvloop==0.19.0 ; sys_platform != 'win32'

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==41.0.7

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Configuration & Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# File handling
aiofiles==23.2.1

# LLM Integration - Simplified versions
langchain==0.1.0
openai==1.6.1
anthropic==0.8.1
tiktoken==0.5.2

# Remove these for now - they cause conflicts
# langchain-openai==0.0.5
# langchain-anthropic==0.0.1

# File Processing
python-docx==1.1.0
python-pptx==0.6.23
markdown==3.5.1
PyPDF2==3.0.1

# Email functionality
fastapi-mail==1.4.1

# HTTP client for external APIs
httpx==0.25.2
requests==2.31.0

# Typing extensions for older Python versions
typing-extensions==4.9.0